\name{quarkW}
\alias{quarkW}
\title{
  Print \pkg{quark}'s warranty statement.
}
\description{
  Print the sections of the GPL-3 that describe the warranty (or
  complete lack thereof) for \pkg{quark}.
}
\usage{
quarkW()
}
\value{
  Text giving the warranty-specific sections of the GPL-3.
}
\author{
  Kyle M. Lang
}
\examples{
## Check quark's warranty:
quarkW()
}